+++ src/hooks/queries/index.ts
@@ -1,3 +1,4 @@
 import useFetchArticle from "./useFetchArticle";
+import useCreateArticle from "./useCreateArticle";
 
-export { useFetchArticle };
+export { useFetchArticle, useCreateArticle };
