+++ src/hooks/queries/useCreateArticle.ts
@@ -1,0 +1,32 @@
+import { useNavigate } from "react-router-dom";
+import { useRecoilState } from "recoil";
+
+import { ArticleStorageState } from "store/persist";
+import { ArticleFormValues } from "types/article";
+import { pseudoLatency } from "./pseudoLatency";
+
+const useCreateArticle = () => {
+  const [articleStorage, setArticleStorage] =
+    useRecoilState(ArticleStorageState);
+
+  const navigate = useNavigate();
+
+  const createArticle = async (data: ArticleFormValues) => {
+    const id =
+      Number(
+        Object.keys(articleStorage)[Object.keys(articleStorage).length - 1],
+      ) + 1;
+
+    await pseudoLatency().then(() => {
+      setArticleStorage({
+        ...(typeof articleStorage === "object" ? articleStorage : {}),
+        [id]: { title: data.title, content: data.content },
+      });
+      navigate(`/article/detail?id=${id}`);
+    });
+  };
+
+  return { createArticle };
+};
+
+export default useCreateArticle;
